design_patterns/SCREENPLAY/requirements.txt
playwright==1.48.0
pytest==8.3.3
pytest-playwright==0.5.2
pytest-xdist==3.6.1
allure-pytest==2.13.5
pytest-bdd==8.1.0
